criação da página de acesso dos clientes e a possibilidade de editar um acesso do cliente

+++ src/main/java/br/com/g4flex/service/ControlOnDutyService.java
@@ -56,7 +56,7 @@ public class ControlOnDutyService {
 	}
 	
 	public int getQuantityPage(int quantity, int total) {
-	 	return (int)Math.ceil(total/quantity);
+	 	return (int)Math.ceil((double)total/quantity);
 	}
 	
 	public List<Object[]> getArrayOfArrayObject(FiltersDTO filters) {
