+++ src/main/java/br/com/g4flex/service/ExtraActivityService.java
@@ -58,7 +58,7 @@ public class ExtraActivityService {
 	}
 	
 	public int getQuantityPage(int quantity, int total) {
-	 	return (int)Math.ceil(total/quantity);
+	 	return (int)Math.ceil((double)total/quantity);
 	}
