+++ src/main/java/br/com/g4flex/service/PointService.java
@@ -65,7 +65,7 @@ public class PointService {
 	}
 	
 	public int getQuantityPage(int quantity, int total) {
-	 	return (int)Math.ceil(total/quantity);
+	 	return (int)Math.ceil((double)total/quantity);
 	}
 	
 	public List<Object[]> getArrayOfArrayObject(FiltersDTO filters) {
